Fix some odd enum issues, possible circular imports
