+++ frontend/src/actions/MeshActions.ts
@@ -2,7 +2,7 @@ import { ActionType, createStandardAction } from 'typesafe-actions';
 import { ActionKeys } from './ActionKeys';
 import { TimeInMilliseconds } from 'types/Common';
 import { MeshDefinition, MeshTarget } from 'types/Mesh';
-import { MeshLayout } from 'pages/Mesh/Mesh';
+import { MeshLayout } from 'pages/Mesh/layouts/layoutFactory';
 
 export const MeshActions = {
   setDefinition: createStandardAction(ActionKeys.MESH_SET_DEFINITION)<MeshDefinition>(),
