+++ frontend/src/pages/Mesh/toolbar/__tests__/MeshFind.test.tsx
@@ -1,10 +1,11 @@
 import * as React from 'react';
 import { shallow } from 'enzyme';
 import { MeshFindComponent } from '../MeshFind';
-import { MeshLayout } from 'pages/Mesh/Mesh';
+import { MeshLayout } from 'pages/Mesh/layouts/layoutFactory';
 
 const testHandler = () => undefined;
 const testSetter = _val => undefined;
+const layout = MeshLayout.Dagre;
 
 // TODO Find out why typescript is unhappy and get rid of all of these ts-ignores
 describe('Parse find value test', () => {
@@ -19,7 +20,7 @@ describe('Parse find value test', () => {
         setFindValue={testSetter}
         setHideValue={testSetter}
         toggleFindHelp={testHandler}
-        layout={MeshLayout.Dagre}
+        layout={layout}
       />
     );
