+++ frontend/src/store/Store.ts
@@ -28,7 +28,7 @@ import { MetricsStatsState } from 'reducers/MetricsStatsState';
 import { CertsInfo } from 'types/CertsInfo';
 import { MeshCluster, MeshDefinition, MeshTarget } from '../types/Mesh';
 import { GraphLayout } from 'pages/Graph/GraphPF';
-import { MeshLayout } from 'pages/Mesh/Mesh';
+import { MeshLayout } from 'pages/Mesh/layouts/layoutFactory';
 
 // Store is the Redux Data store
